03- Mastering OpenAI APIs/01-SentimentAnalysis/requirements.txt
openai
python-dotenv
fastapi==0.95.1
uvicorn==0.20.0
python-multipart